requirements.txt
discord.py
fastapi
uvicorn